+++ frontend/src/app/components/docs/api-docs-nav.component.scss
@@ -6,7 +6,7 @@ p {
 }
 
 p:first-child {
-  margin-top: 0
+  margin-top: 0;
 }
 
 a {
