+++ frontend/src/app/components/docs/api-docs-nav.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit, Input } from '@angular/core';
+import { restApiDocsData } from './api-docs-data';
 
 @Component({
   selector: 'app-api-docs-nav',
@@ -9,10 +10,12 @@ export class ApiDocsNavComponent implements OnInit {
 
   @Input() network: any;
   @Input() collapseItem: any = { toggle: () => {} };
+  restDocs: any[];
 
   constructor() { }
 
   ngOnInit(): void {
+    this.restDocs = restApiDocsData;
   }
 
 }
